+++ GdeIzaci/Models/DTO/ReviewDTO.cs
@@ -5,7 +5,7 @@ namespace GdeIzaci.Models.DTO
     public class ReviewDTO
     {
         public Guid ReviewID { get; set; }
-        public int numberOfStars { get; set; }
+        public double numberOfStars { get; set; }
         public Guid UserID { get; set; }
         public Guid PlaceID { get; set; }
     }
