+++ GdeIzaci/Models/DTO/UpdateReviewDto.cs
@@ -5,7 +5,7 @@ namespace GdeIzaci.Models.DTO
     public class UpdateReviewDto
     {
         [Required]
-        public int NumberOfStars { get; set; }
+        public double NumberOfStars { get; set; }
         [Required]
         public Guid PlaceID { get; set; }
     }
